Get the mcts tests working (note that tictactoe is still not working). Also add a test to connectfour.
